Move non-acceptance tests into separate test file
